% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_image_fetch.R
\name{apply_coord_cartesian}
\alias{apply_coord_cartesian}
\title{Apply \code{coord_cartesian()} to stored plots}
\usage{
apply_coord_cartesian(img_data)
}
\arguments{
\item{img_data}{A \linkS4class{TiffPlotData} object.}
}
\value{
The modified \linkS4class{TiffPlotData} object.
}
\description{
Applies \code{ggplot2::coord_cartesian()} to each plot stored in a
\linkS4class{TiffPlotData} object using its rectangle extent.
}
\examples{
\dontrun{
obj <- fetchTiffData(exampleTiff(), TiffRect(900, 2300, 1400, 2800))
obj <- apply_coord_cartesian(obj)
}
}
