% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R
\name{c,TiffEllipse-method}
\alias{c,TiffEllipse-method}
\title{Combine `TiffEllipse` objects}
\usage{
\S4method{c}{TiffEllipse}(x, ...)
}
\arguments{
\item{...}{one or more `TiffEllipse` objects}
}
\value{
a single `TiffEllipse` with all rows combined
}
\description{
Combine `TiffEllipse` objects
}
