% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPolygon.R
\name{c,TiffPolygon-method}
\alias{c,TiffPolygon-method}
\title{Combine `TiffPolygon` objects}
\usage{
\S4method{c}{TiffPolygon}(x, ...)
}
\arguments{
\item{...}{one or more `TiffPolygon` objects}
}
\value{
a single `TiffPolygon` with all rows combined
}
\description{
Combine `TiffPolygon` objects
}
