% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{c,TiffRect-method}
\alias{c,TiffRect-method}
\title{Combine \code{\link{TiffRect}} objects}
\usage{
\S4method{c}{TiffRect}(x, ...)
}
\arguments{
\item{...}{one or more `TiffRect` objects}
}
\value{
a single `TiffRect` with all rows combined
}
\description{
Concatenates the coordinate rows of two or more `TiffRect` objects.
Duplicate names are made unique automatically (same rule as in `TiffRect()`).
}
\examples{
r1 <- TiffRect(0, 5, 0, 5, name = "a")
r2 <- TiffRect(10, 15, 10, 15, name = "b")
r3 <- c(r1, r2)
}
