% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotData.R
\name{$<-,TiffPlotData-method}
\alias{$<-,TiffPlotData-method}
\title{$<- assignment accessor for TiffPlotData
Set slots or data.frame columns}
\usage{
\S4method{$}{TiffPlotData}(x, name) <- value
}
\arguments{
\item{x}{A TiffPlotData object}

\item{name}{Name of slot or data.frame column}

\item{value}{Value to assign}
}
\description{
$<- assignment accessor for TiffPlotData
Set slots or data.frame columns
}
