% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotData.R
\name{$,TiffPlotData-method}
\alias{$,TiffPlotData-method}
\title{$ accessor for TiffPlotData
Access slots or data.frame columns}
\usage{
\S4method{$}{TiffPlotData}(x, name)
}
\arguments{
\item{x}{A TiffPlotData object}

\item{name}{Name of slot or data.frame column}
}
\description{
$ accessor for TiffPlotData
Access slots or data.frame columns
}
