% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_image_fetch.R
\name{convertTidyToRGB}
\alias{convertTidyToRGB}
\title{Title}
\usage{
convertTidyToRGB(
  img_df,
  red_channel = 1,
  green_channel = 2,
  blue_channel = 3,
  value_var = "norm_value"
)
}
\arguments{
\item{blue_channel}{}
}
\description{
Title
}
