% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_cores.R
\name{.free_mem_macos}
\alias{.free_mem_macos}
\title{macOS: free + inactive + speculative pages from vm_stat}
\usage{
.free_mem_macos()
}
\description{
macOS: free + inactive + speculative pages from vm_stat
}
