% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_cores.R
\name{.free_mem_windows}
\alias{.free_mem_windows}
\title{Windows: FreePhysicalMemory from wmic (kB -> MB), PowerShell fallback}
\usage{
.free_mem_windows()
}
\description{
Windows: FreePhysicalMemory from wmic (kB -> MB), PowerShell fallback
}
