% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_image_fetch.R
\name{exampleTiff}
\alias{exampleTiff}
\title{Path to the example tiff file for testing and examples.}
\usage{
exampleTiff()
}
\value{
Path to the example tiff file
}
\description{
Path to the example tiff file for testing and examples.
}
\examples{
exampleTiff()
}
