% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers_methods.R
\name{ggplot_add.TiffEllipse}
\alias{ggplot_add.TiffEllipse}
\title{Add a `TiffEllipse` to a ggplot using `+`}
\usage{
\method{ggplot_add}{TiffEllipse}(object, plot, object_name)
}
\arguments{
\item{object}{A `TiffEllipse` object.}

\item{plot}{A ggplot object.}

\item{object_name}{Unused; required by `ggplot_add`.}
}
\value{
ggplot object with ellipse annotation added.
}
\description{
Enables syntax like `ggplot(...) + my_ellipse`.
}
