% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers_methods.R
\name{ggplot_add.TiffPolygon}
\alias{ggplot_add.TiffPolygon}
\title{Add a `TiffPolygon` to a ggplot using `+`}
\usage{
\method{ggplot_add}{TiffPolygon}(object, plot, object_name)
}
\arguments{
\item{object}{A `TiffPolygon` object.}

\item{plot}{A ggplot object.}

\item{object_name}{Unused; required by `ggplot_add`.}
}
\value{
ggplot object with polygon annotation added.
}
\description{
Enables syntax like `ggplot(...) + my_polygon`.
}
