% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{ggplot_add.TiffRect}
\alias{ggplot_add.TiffRect}
\title{Add a `TiffRect` to a ggplot using `+`}
\usage{
\method{ggplot_add}{TiffRect}(object, plot, object_name)
}
\arguments{
\item{object}{A `TiffRect` object.}

\item{plot}{A ggplot object.}

\item{object_name}{Unused; required by `ggplot_add`.}
}
\value{
ggplot object with rectangle annotation added.
}
\description{
Enables syntax like `ggplot(...) + my_rect`.
}
