% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers_methods.R
\name{ggplot_add.TiffShapeLayer}
\alias{ggplot_add.TiffShapeLayer}
\title{Add a `TiffShapeLayer` to a ggplot using `+`}
\usage{
\method{ggplot_add}{TiffShapeLayer}(object, plot, object_name)
}
\arguments{
\item{object}{A `TiffShapeLayer` object created by `shape_layer()`.}

\item{plot}{A ggplot object.}

\item{object_name}{Unused; required by `ggplot_add`.}
}
\value{
ggplot object with shape annotation added.
}
\description{
Add a `TiffShapeLayer` to a ggplot using `+`
}
