% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_helpers_methods.R
\name{ggplot_add.TiffShapeNameLayer}
\alias{ggplot_add.TiffShapeNameLayer}
\title{Add a `TiffShapeNameLayer` to a ggplot using `+`}
\usage{
\method{ggplot_add}{TiffShapeNameLayer}(object, plot, object_name)
}
\arguments{
\item{object}{A `TiffShapeNameLayer` object created by `shape_name_layer()`.}

\item{plot}{A ggplot object.}

\item{object_name}{Unused; required by `ggplot_add`.}
}
\value{
ggplot object with shape names added.
}
\description{
Add a `TiffShapeNameLayer` to a ggplot using `+`
}
