% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotData.R
\name{plot,TiffPlotData,ANY-method}
\alias{plot,TiffPlotData,ANY-method}
\title{Plot method for TiffPlotData}
\usage{
\S4method{plot}{TiffPlotData,ANY}(x, y = 1, ...)
}
\arguments{
\item{x}{A TiffPlotData object}

\item{y}{Unused}
}
\description{
Plot method for TiffPlotData
}
