% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_image_fetch.R
\name{read_tiff_meta_data}
\alias{read_tiff_meta_data}
\title{Read TIFF image metadata}
\usage{
read_tiff_meta_data(tiff_path)
}
\arguments{
\item{tiff_path}{Path to the TIFF image file}
}
\value{
A data frame with columns:
  \item{sizeX}{Image width in pixels}
  \item{sizeY}{Image height in pixels}
  \item{sizeC}{Number of channels}
  \item{resolutionLevel}{Resolution level number}
}
\description{
Extracts metadata information from a TIFF image file including image
dimensions and available resolution levels.
}
\examples{
\dontrun{
  img_info <- read_tiff_meta_data("image.tiff")
  print(img_info)
}

}
