% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{rect_intersect}
\alias{rect_intersect}
\title{Intersect two TiffRect objects}
\usage{
rect_intersect(rect, other, invert = FALSE, name = NULL)
}
\arguments{
\item{rect}{TiffRect}

\item{other}{TiffRect}

\item{invert}{logical; when FALSE (default), returns the intersection rectangle
or NULL when there is no overlap. When TRUE, returns whether rectangles do
not overlap.}

\item{name}{optional name(s) for returned intersection rectangle(s)}
}
\value{
A `TiffRect`, `NULL`, or logical depending on `invert`
}
\description{
Intersect two TiffRect objects
}
