% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{rect_resize_abs}
\alias{rect_resize_abs}
\title{Resize a TiffRect to absolute width/height (anchor=center)}
\usage{
rect_resize_abs(rect, width, height, anchor = "center")
}
\arguments{
\item{rect}{TiffRect}

\item{width}{desired width (numeric)}

\item{height}{desired height (numeric)}

\item{anchor}{where to keep fixed: "center" (default), "topleft", "topright", "botleft", "botright"}
}
\value{
resized TiffRect
}
\description{
Resize a TiffRect to absolute width/height (anchor=center)
}
