% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{rect_resize_mult}
\alias{rect_resize_mult}
\title{Resize a TiffRect by multipliers}
\usage{
rect_resize_mult(rect, fx = 1, fy = NULL, anchor = "center")
}
\arguments{
\item{rect}{TiffRect}

\item{fx}{multiplier for width (numeric)}

\item{fy}{multiplier for height (numeric, defaults to fx)}

\item{anchor}{where to keep fixed: "center" (default), "topleft", "topright", "botleft", "botright"}
}
\value{
resized TiffRect
}
\description{
Resize a TiffRect by multipliers
}
