% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{resize_rect_mult}
\alias{resize_rect_mult}
\title{Resize a TiffRect by multipliers}
\usage{
resize_rect_mult(rect, fx = 1, fy = NULL)
}
\arguments{
\item{rect}{TiffRect}

\item{fx}{multiplier for width (numeric)}

\item{fy}{multiplier for height (numeric, defaults to fx)}
}
\value{
resized TiffRect
}
\description{
Resize a TiffRect by multipliers
}
