% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generics.R, R/shape_helpers_methods.R,
%   R/class_TiffRect.R
\name{shape_annotate}
\alias{shape_annotate}
\alias{shape_annotate,ANY,ANY-method}
\alias{shape_annotate,ANY,TiffEllipse-method}
\alias{shape_annotate,ANY,TiffPolygon-method}
\alias{shape_annotate,ANY,TiffRect-method}
\alias{shape_annotate,TiffPlotData,TiffShape-method}
\title{Annotate ggplot objects with shape geometry}
\usage{
shape_annotate(
  p,
  shape,
  color = "green",
  fill = NA,
  alpha = 0.2,
  annotate_center = FALSE,
  n = 120L,
  ...
)

\S4method{shape_annotate}{ANY,ANY}(
  p,
  shape,
  color = "green",
  fill = NA,
  alpha = 0.2,
  annotate_center = FALSE,
  n = 120L,
  ...
)

\S4method{shape_annotate}{ANY,TiffEllipse}(
  p,
  shape,
  color = "green",
  fill = NA,
  alpha = 0.2,
  annotate_center = FALSE,
  n = 120L,
  ...
)

\S4method{shape_annotate}{ANY,TiffPolygon}(
  p,
  shape,
  color = "green",
  fill = NA,
  alpha = 0.2,
  annotate_center = FALSE,
  n = 120L,
  ...
)

\S4method{shape_annotate}{ANY,TiffRect}(
  p,
  shape,
  color = "green",
  fill = NA,
  alpha = 0.2,
  annotate_center = FALSE,
  n = 120L,
  ...
)

\S4method{shape_annotate}{TiffPlotData,TiffShape}(
  p,
  shape,
  color = "green",
  fill = NA,
  alpha = 0.2,
  annotate_center = FALSE,
  n = 120L,
  ...
)
}
\arguments{
\item{p}{ggplot object.}

\item{shape}{A `TiffShape` object.}

\item{color}{Border/line color.}

\item{fill}{Fill color.}

\item{alpha}{Alpha transparency.}

\item{annotate_center}{If TRUE, annotate center points instead of boundaries.}

\item{n}{Number of vertices used to draw ellipse boundaries.}

\item{...}{Additional arguments passed to ggplot layer calls.}
}
\value{
A ggplot object with added shape annotation layer.

for `p` as `TiffPlotData`, returns a `TiffPlotData` with the active plot annotated
}
\description{
Annotate ggplot objects with shape geometry
}
