% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generics.R, R/shape_helpers_methods.R,
%   R/class_TiffRect.R
\name{shape_center_points}
\alias{shape_center_points}
\alias{shape_center_points,TiffEllipse-method}
\alias{shape_center_points,TiffPolygon-method}
\alias{shape_center_points,ANY-method}
\alias{shape_center_points,TiffRect-method}
\title{Return shape center points}
\usage{
shape_center_points(shape)

\S4method{shape_center_points}{TiffEllipse}(shape)

\S4method{shape_center_points}{TiffPolygon}(shape)

\S4method{shape_center_points}{ANY}(shape)

\S4method{shape_center_points}{TiffRect}(shape)
}
\arguments{
\item{shape}{A `TiffShape` object.}
}
\value{
data.frame with columns `x` and `y`.
}
\description{
Computes one representative center point per shape row.
}
