% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R, R/class_TiffRect.R
\name{shape_resize_abs}
\alias{shape_resize_abs}
\alias{shape_resize_abs,ANY-method}
\alias{shape_resize_abs,TiffEllipse-method}
\alias{shape_resize_abs,TiffPolygon-method}
\alias{shape_resize_abs,TiffRect-method}
\title{Resize shape geometry to absolute dimensions}
\usage{
shape_resize_abs(shape, width, height, anchor = "center")

\S4method{shape_resize_abs}{ANY}(shape, width, height, anchor = "center")

\S4method{shape_resize_abs}{TiffEllipse}(shape, width, height, anchor = "center")

\S4method{shape_resize_abs}{TiffPolygon}(shape, width, height, anchor = "center")

\S4method{shape_resize_abs}{TiffRect}(shape, width, height, anchor = "center")
}
\arguments{
\item{shape}{A `TiffShape` object.}

\item{width}{Target shape width in x units.}

\item{height}{Target shape height in y units.}

\item{anchor}{Anchor point used to keep shape location fixed while resizing.}
}
\value{
Resized shape object.
}
\description{
Resize shape geometry to absolute dimensions
}
