% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R, R/class_TiffRect.R
\name{shape_resize_mult}
\alias{shape_resize_mult}
\alias{shape_resize_mult,TiffShape-method}
\alias{shape_resize_mult,TiffRect-method}
\title{Resize shape geometry by multiplicative scale factors}
\usage{
shape_resize_mult(shape, fx = 1, fy = NULL, anchor = "center")

\S4method{shape_resize_mult}{TiffShape}(shape, fx = 1, fy = NULL, anchor = "center")

\S4method{shape_resize_mult}{TiffRect}(shape, fx = 1, fy = NULL, anchor = "center")
}
\arguments{
\item{shape}{A `TiffShape` object.}

\item{fx}{Width scale factor.}

\item{fy}{Height scale factor (defaults to `fx`).}

\item{anchor}{Anchor point used to keep shape location fixed while resizing.}
}
\value{
Resized shape object.
}
\description{
Resize shape geometry by multiplicative scale factors
}
