% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generics.R, R/shape_helpers_methods.R,
%   R/class_TiffRect.R
\name{shape_shift}
\alias{shape_shift}
\alias{shape_shift,TiffEllipse-method}
\alias{shape_shift,TiffPolygon-method}
\alias{shape_shift,TiffRect-method}
\title{Shift a shape in x/y dimensions}
\usage{
shape_shift(shape, dx = 0, dy = 0)

\S4method{shape_shift}{TiffEllipse}(shape, dx = 0, dy = 0)

\S4method{shape_shift}{TiffPolygon}(shape, dx = 0, dy = 0)

\S4method{shape_shift}{TiffRect}(shape, dx = 0, dy = 0)
}
\arguments{
\item{shape}{A `TiffShape` object (`TiffRect`, `TiffEllipse`, or `TiffPolygon`)}

\item{dx}{shift in x direction (numeric)}

\item{dy}{shift in y direction (numeric)}
}
\value{
shifted object of the same class as input
}
\description{
Shift a shape in x/y dimensions
}
