% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotData.R
\name{show,TiffPlotData-method}
\alias{show,TiffPlotData-method}
\title{Show method for TiffPlotData}
\usage{
\S4method{show}{TiffPlotData}(object)
}
\arguments{
\item{object}{A TiffPlotData object}
}
\description{
Show method for TiffPlotData
}
