% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{show,TiffRect-method}
\alias{show,TiffRect-method}
\title{Show method for TiffRect}
\usage{
\S4method{show}{TiffRect}(object)
}
\arguments{
\item{object}{TiffRect}
}
\description{
Show method for TiffRect
}
