% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{[,TiffRect-method}
\alias{[,TiffRect-method}
\title{Subset a \code{\link{TiffRect}} by row index}
\usage{
\S4method{[}{TiffRect}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{TiffRect object}

\item{i}{integer, logical, or character index}

\item{j}{unused}

\item{drop}{unused}
}
\value{
a `TiffRect` containing only the selected rows
}
\description{
Subset a \code{\link{TiffRect}} by row index
}
\examples{
r <- TiffRect(xmin = 1:4, xmax = 2:5, ymin = 1:4, ymax = 2:5,
              name = c("a", "b", "c", "d"))
r[2:3]          # rows 2 and 3
r[c(TRUE, FALSE, TRUE, FALSE)]  # logical index
}
