% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R
\docType{class}
\name{TiffEllipse-class}
\alias{TiffEllipse-class}
\title{Ellipse definition S4 class}
\description{
Stores ellipses using center coordinates and x/y radii.
}
\section{Slots}{

\describe{
\item{\code{coords}}{data.frame with x0/y0/radius_x/radius_y/name columns}
}}

