% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R
\name{TiffEllipse}
\alias{TiffEllipse}
\alias{$,TiffEllipse-method}
\title{Construct a TiffEllipse}
\usage{
TiffEllipse(x0, y0, radius_x, radius_y = radius_x, name = "ellipse")

\S4method{$}{TiffEllipse}(x, name)
}
\arguments{
\item{x0}{x center coordinate}

\item{y0}{y center coordinate}

\item{radius_x}{ellipse x radius}

\item{radius_y}{ellipse y radius (defaults to radius_x)}

\item{name}{Name of slot to be replaced/retrieved}
}
\value{
a `TiffEllipse` object
}
\description{
Construct a TiffEllipse
}
