% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotData.R
\docType{class}
\name{TiffPlotData-class}
\alias{TiffPlotData-class}
\title{An S4 Class to Store a Data Frame and List of ggplot Objects}
\description{
This class combines a data frame with a list of associated ggplot objects
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame object}

\item{\code{plots}}{A list of ggplot objects (should be named)}

\item{\code{activePlot}}{A character string specifying which plot to display (must be a name in the plots list)}

\item{\code{tiff_path}}{A character scalar giving the path of the TIFF file used to produce the object (may be NA if not applicable)}

\item{\code{resolution}}{A numeric scalar recording the resolution level read from the TIFF (NA when unknown)}

\item{\code{precalc_max}}{A data.frame of precalculated min/max values used for normalization}

\item{\code{rect}}{A \linkS4class{TiffRect} object describing the rectangular region that was plotted}

\item{\code{img_info}}{A data.frame containing TIFF metadata returned by \code{read_tiff_meta_data} (may be empty)}
}}

\examples{
\dontrun{
library(ggplot2)
df <- data.frame(x = 1:10, y = rnorm(10))
p1 <- ggplot(df, aes(x = x, y = y)) + geom_point()
r <- TiffRect(0, 10, 0, 10)
# manual construction using new():
obj <- new("TiffPlotData", data = df, plots = list(scatter=p1), activePlot = "scatter",
           tiff_path = NA_character_, resolution = NA_real_, precalc_max = data.frame(), rect = r,
           img_info = data.frame())
# Preferred convenience constructor:
obj2 <- TiffPlotData(df, plots = list(scatter=p1), activePlot = "scatter")
}

}
