% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotDataMasked.R
\docType{class}
\name{TiffPlotDataMasked-class}
\alias{TiffPlotDataMasked-class}
\title{An S4 class for mask-filtered TIFF signal data}
\description{
Extends \linkS4class{TiffPlotData} by storing decoded mask points and
mask color mappings.
}
\section{Slots}{

\describe{
\item{\code{mask_points}}{A data.frame of decoded mask coordinates and metadata.}

\item{\code{mask_color_mappings}}{A data.frame mapping mask identifiers to colors.}
}}

