% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffPlotDataMasked.R
\name{TiffPlotDataMasked}
\alias{TiffPlotDataMasked}
\title{Constructor for TiffPlotDataMasked objects}
\usage{
TiffPlotDataMasked(
  data,
  plots,
  activePlot = NULL,
  tiff_path = NA_character_,
  resolution = NA_real_,
  precalc_max = data.frame(),
  rect = TiffRect(0, 1, 0, 1),
  img_info = data.frame(),
  mask_points = data.frame(),
  mask_color_mappings = data.frame()
)
}
\arguments{
\item{data}{A data.frame of image signal data.}

\item{plots}{A named list of ggplot objects.}

\item{activePlot}{Active plot name. If NULL, first plot name is used.}

\item{tiff_path}{Character path to TIFF file.}

\item{resolution}{Numeric resolution level.}

\item{precalc_max}{Data frame of min/max normalization values per channel.}

\item{rect}{A \linkS4class{TiffRect} region.}

\item{img_info}{TIFF metadata data.frame.}

\item{mask_points}{Decoded mask points and metadata.}

\item{mask_color_mappings}{Mask ID/text to color mapping table.}
}
\value{
A \linkS4class{TiffPlotDataMasked} object.
}
\description{
Constructor for TiffPlotDataMasked objects
}
