% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R
\docType{class}
\name{TiffPolygon-class}
\alias{TiffPolygon-class}
\title{Polygon definition S4 class}
\description{
Stores polygons as list-columns of x/y vertices.
}
\section{Slots}{

\describe{
\item{\code{coords}}{data.frame with x/y/name columns where x and y are list-columns}
}}

