% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R
\name{TiffPolygon}
\alias{TiffPolygon}
\alias{$,TiffPolygon-method}
\title{Construct a TiffPolygon}
\usage{
TiffPolygon(x, y, name = "polygon")

\S4method{$}{TiffPolygon}(x, name)
}
\arguments{
\item{x}{numeric vector of x vertices or list of numeric vectors}

\item{y}{numeric vector of y vertices or list of numeric vectors}

\item{name}{Name of slot to be replaced/retrieved}
}
\value{
a `TiffPolygon` object
}
\description{
Construct a TiffPolygon
}
