% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffShape.R
\docType{class}
\name{TiffShape-class}
\alias{TiffShape-class}
\title{Base class for TIFF geometry shapes}
\description{
`TiffShape` is a virtual S4 class that stores shared structure for shape
objects used in TIFF coordinate space.
}
\section{Slots}{

\describe{
\item{\code{coords}}{data.frame containing shape coordinates and a `name` column}
}}

