% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{rect_annotate}
\alias{rect_annotate}
\title{Annotate a ggplot with a TiffRect}
\usage{
rect_annotate(p, rect, color = "green", fill = NA, alpha = 0.2, ...)
}
\arguments{
\item{p}{ggplot object}

\item{rect}{TiffRect}

\item{color}{border color (default "red")}

\item{fill}{fill color (default NA)}

\item{alpha}{fill alpha (default 0.2)}

\item{...}{additional args passed to geom_rect}
}
\value{
ggplot object with rectangle layer added
}
\description{
Annotate a ggplot with a TiffRect
}
