% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{rect_shift}
\alias{rect_shift}
\title{Shift a TiffRect by dx and dy}
\usage{
rect_shift(rect, dx = 0, dy = 0)
}
\arguments{
\item{rect}{TiffRect object}

\item{dx}{shift in x direction (numeric)}

\item{dy}{shift in y direction (numeric)}
}
\value{
shifted TiffRect
}
\description{
Shift a TiffRect by dx and dy
}
