% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_A_TiffShape.R, R/class_TiffRect.R
\name{shape_shift}
\alias{shape_shift}
\alias{shape_shift,TiffEllipse-method}
\alias{shape_shift,TiffPolygon-method}
\alias{shape_shift,TiffRect-method}
\title{Shift a shape in x/y dimensions}
\usage{
shape_shift(shape, dx = 0, dy = 0)

\S4method{shape_shift}{TiffEllipse}(shape, dx = 0, dy = 0)

\S4method{shape_shift}{TiffPolygon}(shape, dx = 0, dy = 0)

\S4method{shape_shift}{TiffRect}(shape, dx = 0, dy = 0)
}
\arguments{
\item{shape}{A `TiffShape` object (`TiffRect`, `TiffEllipse`, or `TiffPolygon`)}

\item{dx}{shift in x direction (numeric)}

\item{dy}{shift in y direction (numeric)}
}
\value{
shifted object of the same class as input
}
\description{
Shift a shape in x/y dimensions

#' @rdname TiffRect
#' @param value New value for slot.
setReplaceMethod("$", "TiffRect",
                 function(x, name, value)
                 {
                     warn_msg = "This assignment is not supported.  No effect."
                     switch (name,
                             view_size = {
                                 if(is.null(value)){
                                     value = -1
                                 }
                                 if(is.na(value)){
                                     value = -1
                                 }
                                 x@view_size = value
                             },
                             window_size = {
                                 x@window_size = value
                             },
                             read_mode = {
                                 stopifnot(value %in% c("bam_SE", "bam_PE", "bigwig", "null"))
                                 x@read_mode = value
                             },
                             fetch_options = {
                                 x@fetch_options = value
                             },
                             meta_data = {
                                 x@meta_data = value
                             },
                             warning(warn_msg)
                     )
                     validObject(x)
                     x
                 })
}
