% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\docType{class}
\name{TiffRect-class}
\alias{TiffRect-class}
\title{Rectangle definition S4 class}
\description{
Stores a rectangular region definition with numeric boundaries.
}
\section{Slots}{

\describe{
\item{\code{coords}}{data.frame with xmin/xmax/ymin/ymax/name columns}
}}

\examples{
r <- TiffRect(c(10, 15), 20, 40, 50, name = "roi1")
r2 <- rect_shift(r, dx = 5, dy = -2)
r3 <- rect_resize_mult(r2, fx = 2, )
r4 <- rect_resize_mult(r2, fx = .5)

library(ggplot2)
library(magrittr)
ggplot() \%>\%
  rect_annotate(r, fill = "red", alpha = 0.2) \%>\%
  rect_annotate(r2, fill = "green", alpha = 0.2)

ggplot() \%>\%
  rect_annotate(r, fill = "red", alpha = 0.2) \%>\%
  rect_annotate(r3, fill = "blue", alpha = 0.2) \%>\%
  rect_annotate(r4, fill = "yellow", alpha = 0.2)

}
