% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TiffRect.R
\name{TiffRect}
\alias{TiffRect}
\title{Construct a TiffRect}
\usage{
TiffRect(xmin, xmax, ymin, ymax, name = "rect")
}
\arguments{
\item{xmin}{left x coordinate}

\item{xmax}{right x coordinate}

\item{ymin}{bottom y coordinate}

\item{ymax}{top y coordinate}

\item{name}{optional name(s). Recycled when length 1}
}
\value{
a `TiffRect` object
}
\description{
Construct a TiffRect
}
